Consolidate ogbn_papers_100m_gat, ogbn_products_gat.py and ogbn_products_sage.py into ogbn_train.py
