+++ examples/ogbn_train.py
@@ -7,19 +7,46 @@ import time
 import torch
 import torch.nn.functional as F
 from ogb.nodeproppred import Evaluator, PygNodePropPredDataset
+from torch.nn import Linear as Lin
 from tqdm import tqdm
 
 from torch_geometric.loader import NeighborLoader
-from torch_geometric.nn import SAGEConv
+from torch_geometric.nn import SAGEConv, GATConv
 from torch_geometric.utils import to_undirected
 
 parser = argparse.ArgumentParser()
+parser.add_argument(
+    "--dataset",
+    type=str,
+    default="ogbn-papers100M",
+    choices=["ogbn-papers100M", "ogbn-products"],
+    help="Dataset name.",
+)
+parser.add_argument(
+    "--dataset_dir",
+    type=str,
+    default="/workspace/data",
+    help="Root directory of dataset.",
+)
+parser.add_argument(
+    "--dataset_subdir",
+    type=str,
+    default="ogb-papers100M",
+    help="directory of dataset.",
+)
+parser.add_argument(
+    "--use_gat",
+    action='store_true',
+    help="Wether or not to use graphsage model",
+)
 parser.add_argument('--device', type=str, default='cuda')
 parser.add_argument("-e", "--epochs", type=int, default=10,
                     help='number of training epochs.')
 parser.add_argument("--runs", type=int, default=4, help='number of runs.')
 parser.add_argument("--num_layers", type=int, default=3,
                     help='number of layers.')
+parser.add_argument("--num_heads", type=int, default=2,
+                    help='number of heads for GAT model.')
 parser.add_argument("-b", "--batch_size", type=int, default=1024,
                     help='batch size.')
 parser.add_argument("--workers", type=int, default=12,
@@ -53,13 +80,13 @@ neighbors = args.neighbors.split(',')
 num_neighbors = [int(i) for i in neighbors]
 log_interval = args.log_interval
 
-root = osp.join(osp.dirname(osp.realpath(__file__)), '..', 'data',
-                'products')
+root = osp.join(args.dataset_dir, args.dataset_subdir)
 print("The root is: ", root)
-dataset = PygNodePropPredDataset('ogbn-products', root)
+dataset = PygNodePropPredDataset(name=args.dataset, root=root)
 split_idx = dataset.get_idx_split()
-evaluator = Evaluator(name='ogbn-products')
-data = dataset[0]
+evaluator = Evaluator(name=args.dataset)
+
+data=dataset[0]
 if not args.use_directed_graph:
     start_undirected = time.time()
     print("use undirected graph")
@@ -159,9 +186,80 @@ class SAGE(torch.nn.Module):
 
         return x_all
 
+class GAT(torch.nn.Module):
+    def __init__(self, in_channels, hidden_channels, out_channels, num_layers, dropout,
+                 heads):
+        super().__init__()
 
-model = SAGE(dataset.num_features, num_hidden_channels, dataset.num_classes,
+        self.num_layers = num_layers
+
+        self.convs = torch.nn.ModuleList()
+        self.convs.append(GATConv(in_channels, hidden_channels,
+                                  heads))
+        for _ in range(num_layers - 2):
+            self.convs.append(
+                GATConv(heads * hidden_channels, hidden_channels, heads))
+        self.convs.append(
+            GATConv(heads * hidden_channels, out_channels, heads,
+                    concat=False))
+
+        self.skips = torch.nn.ModuleList()
+        self.skips.append(Lin(in_channels, hidden_channels * heads))
+        for _ in range(num_layers - 2):
+            self.skips.append(
+                Lin(hidden_channels * heads, hidden_channels * heads))
+        self.skips.append(Lin(hidden_channels * heads, out_channels))
+        self.dropout = dropout
+
+    def reset_parameters(self):
+        for conv in self.convs:
+            conv.reset_parameters()
+        for skip in self.skips:
+            skip.reset_parameters()
+
+    def forward(self, x, edge_index):
+        for i, (conv, skip) in enumerate(zip(self.convs, self.skips)):
+            x = conv(x, edge_index) + skip(x)
+            if i != self.num_layers - 1:
+                x = F.elu(x)
+                x = F.dropout(x, p=self.dropout, training=self.training)
+        return x
+
+    def inference(self, x_all):
+        pbar = tqdm(total=x_all.size(0) * self.num_layers)
+        pbar.set_description('Evaluating')
+
+        # Compute representations of nodes layer by layer, using *all*
+        # available edges. This leads to faster computation in contrast to
+        # immediately computing the final representations of each batch.
+        for i in range(self.num_layers):
+            xs = []
+            for batch in subgraph_loader:
+                x = x_all[batch.n_id].to(device)
+                edge_index = batch.edge_index.to(device)
+                x = self.convs[i](x, edge_index) + self.skips[i](x)
+                x = x[:batch.batch_size]
+                if i != self.num_layers - 1:
+                    x = F.elu(x)
+                xs.append(x.cpu())
+
+                pbar.update(batch.batch_size)
+
+            x_all = torch.cat(xs, dim=0)
+
+        pbar.close()
+
+        return x_all
+
+if args.use_gat:
+    print("GAT model")
+    model = GAT(dataset.num_features, num_hidden_channels, dataset.num_classes, 
+            num_layers=num_layers, dropout=args.dropout, heads=args.num_heads)
+else:
+    print("GraphSage model")
+    model = SAGE(dataset.num_features, num_hidden_channels, dataset.num_classes,
              num_layers=num_layers, dropout=args.dropout)
+
 model = model.to(device)
